# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "在庫管理を行うAPIサーバー",
        "title": "inventory_control API",
        "contact": {
            "name": "API Support",
            "email": "connect0459@github.com"
        },
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/api/v1",
    "paths": {
...
        "/types_genres": {
            "get": {
                "description": "全てのジャンル情報を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types_genres"
                ],
                "summary": "ジャンルのリストを取得する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TypeGenre"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "新しいジャンル情報の作成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types_genres"
                ],
                "summary": "新しいジャンルを作成する",
                "parameters": [
                    {
                        "description": "ジャンル情報",
                        "name": "TypeGenre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TypeGenre"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.TypeGenre"
                        }
                    }
                }
            }
        },
        "/types_genres/{id}": {
            "get": {
                "description": "IDによるジャンル情報の取得",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types_genres"
                ],
                "summary": "特定のIDでジャンルを取得する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ジャンルのID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TypeGenre"
                        }
                    }
                }
            },
            "put": {
                "description": "IDによるジャンル情報の更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types_genres"
                ],
                "summary": "特定のIDでジャンルを更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ジャンルのID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ジャンル情報",
                        "name": "TypeGenre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TypeGenre"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.TypeGenre"
                        }
                    }
                }
            },
            "delete": {
                "description": "IDによるジャンル情報の削除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types_genres"
                ],
                "summary": "特定のIDでジャンルを削除する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ジャンルのID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ジャンル情報が削除されました",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.TypeGenre": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "レコードの作成日時\nin: body\nrequired: true",
                    "type": "string",
                    "example": "2023-06-20T21:20:14.000000Z"
                },
                "deleted_at": {
                    "description": "レコードの削除日時\nin: body\nrequired: false",
                    "type": "string"
                },
                "id": {
                    "description": "ジャンルID\nin: body\nrequired: true",
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "description": "ジャンル名\nin: body\nrequired: true",
                    "type": "string",
                    "example": "情報学"
                },
                "updated_at": {
                    "description": "レコードの更新日時\nin: body\nrequired: true",
                    "type": "string",
                    "example": "2023-06-20T21:20:14.000000Z"
                }
            }
        }
    }
}